+++ src/main/java/io/github/flibio/minigamecore/events/ArenaStateChangeEvent.java
@@ -22,12 +22,39 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package io.github.minigamecore.teams;
+package io.github.flibio.minigamecore.events;
 
-public class TeamManager {
+import io.github.minigamecore.arena.Arena;
+import org.spongepowered.api.event.cause.Cause;
+import org.spongepowered.api.event.cause.NamedCause;
+import org.spongepowered.api.event.impl.AbstractEvent;
 
-    public TeamManager() {
-        // TODO
+public class ArenaStateChangeEvent extends AbstractEvent {
+
+    private Arena arena;
+    private Object plugin;
+
+    /**
+     * Called when an arena changes its state.
+     * 
+     * @param arena The arena whose state changed.
+     */
+    public ArenaStateChangeEvent(Arena arena, Object plugin) {
+        this.arena = arena;
+    }
+
+    /**
+     * Gets whichever arena's state changed.
+     * 
+     * @return The arena whose state changed.
+     */
+    public Arena getArena() {
+        return this.arena;
+    }
+
+    @Override
+    public Cause getCause() {
+        return Cause.of(NamedCause.of("MinigameCore", plugin));
     }
 
 }
